+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Change Log
 ## Staged
 #### Features
 * adiciona deleção das checklists (PL-3217)
+* Cria funcionalidade de insert de processo de acoplamento (PL-3210)
 
 ### Refactors
 * Realiza modificação mo CRUD de veiculos, incluindo os campos ```motorizado``` e ```possuiHubodometro``` (PL-3223)